# goldtag-api/swaggerDoc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Goldtag App",
    "contact": {
      "name": "Goldtag App",
      "url": "https://goldtag.org",
      "email": "info@goldtag.net"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "variables": {}
    }
  ],
  "paths": {
...
    "/get/{id}&{active}": {
      "get": {
        "tags": ["Content"],
        "summary": "get content",
        "description": "gets filtered content from database",
        "operationId": "getcontent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "all"
            }
          },
          {
            "name": "active",
            "in": "path",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "all"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/successresponse"
                    },
                    {
                      "example": {
                        "success": true,
                        "message": "Fetching contents successfull ",
                        "data": [
                          {
                            "category": "Features",
                            "title": "Features of Test App",
                            "feature1": {
                              "title": "Sharing",
                              "description": "Sharing is caring",
                              "media": "Photo"
                            }
                          }
                        ]
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/badrequest"
                }
              }
            }
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "successresponse": {
        "title": "success",
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "badrequest": {
        "title": "badrequest",
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          }
        },
        "example": {
          "success": false,
          "error": "Bad request please type in valid values/parameters"
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}